+++ lib/src/foundation.dart
@@ -1,0 +1,4 @@
+const foundationSource = '''
+library foundation;
+export 'package:flutter/src/foundation/change_notifier.dart';
+''';