+++ lib/src/material.dart
@@ -7,4 +7,6 @@ export 'package:flutter/src/material/app_bar.dart';
 export 'package:flutter/src/material/colors.dart';
 export 'package:flutter/src/material/floating_action_button.dart';
 export 'package:flutter/src/material/scaffold.dart';
+export 'package:flutter/src/material/snack_bar.dart';
+export 'package:flutter/src/material/text_field.dart';
 ''';