+++ lib/src/widgets.dart
@@ -5,6 +5,7 @@ export 'package:flutter/foundation.dart' show UniqueKey;
 export 'package:flutter/src/widgets/app.dart';
 export 'package:flutter/src/widgets/basic.dart';
 export 'package:flutter/src/widgets/container.dart';
+export 'package:flutter/src/widgets/editable_text.dart';
 export 'package:flutter/src/widgets/framework.dart';
 export 'package:flutter/src/widgets/text.dart';
 export 'package:flutter/src/widgets/navigator.dart';