+++ lib/src/widgets/text.dart
@@ -2,6 +2,7 @@ import 'package:dart_eval/dart_eval.dart';
 import 'package:dart_eval/dart_eval_bridge.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
+import 'package:flutter_eval/src/painting/text_style.dart';
 import 'package:flutter_eval/src/widgets/framework.dart';
 
 class $Text implements Text, $Instance {
@@ -10,7 +11,9 @@ class $Text implements Text, $Instance {
   static const $declaration = BridgeClassDef(BridgeClassType($type, $extends: $StatelessWidget$bridge.$type),
       constructors: {
         '': BridgeConstructorDef(BridgeFunctionDef(returns: BridgeTypeAnnotation($type), params: [
-          BridgeParameter('data', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.stringType)), false)
+          BridgeParameter('data', BridgeTypeAnnotation(BridgeTypeRef.type(RuntimeTypes.stringType)), false),
+        ], namedParams: [
+          BridgeParameter('style', BridgeTypeAnnotation($TextStyle.$type, nullable: true), true),
         ]))
       },
       methods: {},
@@ -22,7 +25,10 @@ class $Text implements Text, $Instance {
   $Text.wrap(this.$value);
 
   static $Value? $new(Runtime runtime, $Value? target, List<$Value?> args) {
-    return $Text.wrap(Text(args[0]!.$value));
+    return $Text.wrap(Text(
+      args[0]!.$value,
+      style: args[1]?.$value,
+    ));
   }
 
   @override