+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.4.0
+## 0.4.1
+- Support for SnackBar, ScaffoldMessenger, TextField, TextStyle, FontWeight, 
+  TextEditingController, and ChangeNotifier
+- Improve support for other widgets
+- Fix a bug in Column causing an error if you didn't specify its
+  MainAxisAlignment
 
+## 0.4.0
 - Initial public release